SEQUOIADBMAINSTREAM-9632:sdb 安装时创建 /var/sequoiadb

+++ script/service_control.sh
@@ -8,12 +8,18 @@ EXEC_PATH=""
 DO_FORCE_STOP_SERVICE=0
 DO_START_SERVICE=0
 DO_STOP_SERVICE=0
+DO_CREATE_DIR=0
+USER_NAME=""
+GROUP_NAME=""
 AUTO_START=""
 
 function stopService()
 {
    # InstallBuild will copy sdbcmtop/sdbstop into /tmp directory,
    # make sure it have the right permission
+   if [ "$EXEC_PATH" == "" ]; then
+      exit 1
+   fi
    chmod 0755 $EXEC_PATH/sdbcmtop
    chmod 0755 $EXEC_PATH/sdbstop
    chmod 0755 $EXEC_PATH/stpstop
@@ -27,6 +33,9 @@ function stopService()
 
 function forceStopService()
 {
+   if [ "$EXEC_PATH" == "" ]; then
+      exit 1
+   fi
    service sdbcm stop
    $EXEC_PATH/sdblist -t all | grep -v "Total" | awk '{print $2}' | awk -F '(' '{print $2}' | awk -F ')' '{print "kill -9 " $1}' | bash
    nodeNum=`$EXEC_PATH/sdblist -t all | grep -v "Total" | wc -l`
@@ -35,14 +44,35 @@ function forceStopService()
 
 function startService()
 {
+   if [ "$EXEC_PATH" == "" ]; then
+      exit 1
+   fi
    service sdbcm start
    # return 127 when didnot have node, so, donot check return code
    $EXEC_PATH/sdbstart -t all > /dev/null 2>&1
 }
 
+function createDir()
+{
+   # check userName and groupName is empty
+   if [ "$USER_NAME" == "" ] || [ "$GROUP_NAME" == "" ]; then
+      exit 1
+   fi
+
+   # if /var/sequoiadb not exist, create it
+   if [ ! -d /var/sequoiadb ]; then
+      mkdir -p /var/sequoiadb
+   fi
+
+   chown $USER_NAME:$GROUP_NAME /var/sequoiadb
+   test $? -ne 0 && exit $?
+   chmod 777 /var/sequoiadb
+   test $? -ne 0 && exit $?
+}
+
 main()
 {
-   ARGS=`getopt -o h --long path:,cmPort:,startService:,stopService:,forceStopService: -n 'service_control.sh' -- "$@"`
+   ARGS=`getopt -o h --long path:,cmPort:,startService:,stopService:,forceStopService:,createDir,userName:,groupName: -n 'service_control.sh' -- "$@"`
    ret=$?
    test $ret -ne 0 && return $ret
    eval set -- "${ARGS}"
@@ -62,6 +92,15 @@ main()
          --forceStopService ) DO_FORCE_STOP_SERVICE=$2
                               shift 2
                               ;;
+         --createDir )        DO_CREATE_DIR=1
+                              shift 1
+                              ;;
+         --userName )         USER_NAME=$2
+                              shift 2
+                              ;;
+         --groupName )        GROUP_NAME=$2
+                              shift 2
+                              ;;
          -- )                 shift
                               break
                               ;;
@@ -71,9 +110,6 @@ main()
       esac
    done
 
-   if [ "$EXEC_PATH" == "" ]; then
-      exit 1
-   fi
 
    if [ "$DO_START_SERVICE" == "true" ]; then
       startService
@@ -87,6 +123,11 @@ main()
    
    if [ "$DO_STOP_SERVICE" == "true" ]; then
       stopService
+      exit 0
+   fi
+
+   if [ "$DO_CREATE_DIR" == "1" ]; then
+      createDir
       exit 0
    fi
 }
